requirements.txt
requests>=2.31.0
ipwhois>=1.2.0
fpdf2>=2.7.7
rich>=13.7.0
python-dotenv>=1.0.0